Fixed some implicit conversion warnings for decimal/integer/dyadic

+++ source/numeric/integer.hpp
@@ -59,7 +59,7 @@ class Nat64 {
   public:
     Nat64() : _m(0u) { }
     template<class M, EnableIf<And<IsBuiltinIntegral<M>,IsBuiltinUnsigned<M>>> = dummy> Nat64(M m) : _m(m) { assert(_m==m); }
-    template<class N, EnableIf<And<IsBuiltinIntegral<N>,IsBuiltinSigned<N>>> = dummy> Nat64(N n) : _m(n) { assert(n>=0); assert((int64_t)_m==n);
+    template<class N, EnableIf<And<IsBuiltinIntegral<N>,IsBuiltinSigned<N>>> = dummy> Nat64(N n) : _m(static_cast<uint64_t>(n)) { assert(n>=0); assert((int64_t)_m==n);
         assert(uint64_t(int64_t(_m))==_m); }
     uint64_t get_ui() const { return _m; }
 };
@@ -68,7 +68,7 @@ class Int32 {
     int32_t _n;
   public:
     Int32() : _n(0) { }
-    template<class M, EnableIf<And<IsBuiltinIntegral<M>,IsBuiltinUnsigned<M>>> = dummy> Int32(M m) : _n(m) { assert(_n>=0); assert((uint32_t)_n==m); }
+    template<class M, EnableIf<And<IsBuiltinIntegral<M>,IsBuiltinUnsigned<M>>> = dummy> Int32(M m) : _n(static_cast<int32_t>(m)) { assert(_n>=0); assert((uint32_t)_n==m); }
     template<class N, EnableIf<And<IsBuiltinIntegral<N>,IsBuiltinSigned<N>>> = dummy> Int32(N n) : _n(n) { assert(_n==n); }
     int32_t get_si() const { return _n; }
 };
